requirements.txt
DateTime==4.3
Flask==1.1.2
Flask-API==2.0
Flask-Login==0.4.1
gc-python-utils==0.0.1
gevent==20.5.0
gevent-websocket==0.10.1
gunicorn==20.0.4
Jinja2==2.9.6
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.4.1
numba==0.53.1
numpy==1.19.2
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==4.2.1
psycopg2==2.8.4
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pytz==2019.3
pytzdata==2019.3
requests==2.23.0
scipy==1.6.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.5.0rc1
tensorflow-estimator==2.3.0
Werkzeug==1.0.1
